Replace helpers with new schema builder mechanism
- Create DefinitionsSchema class, which accepts definitions and returns
  all typeDefs and resolvers

- Generate schemas and resolvers from:
  - Each resource with appropriate definitions
  - Legacy schema.graphql and legacy workspace resolvers (now moved)
  - Standard-form Query and Mutation for each resource

- Create abstract DefinitionSchema class for a single definition

- Create concrete DefinitionSchema classes for mutable vs. immutable
  resources, since they have different mutation names
