+++ packages/db/src/workspace/graphql/index.ts
@@ -1,1 +1,2 @@
 export * from "./types";
+export * from "./schema";
